Move draggable code into it's own component
